### 2025-04-07 cf0faea
Add JOSS badge
