+++ paper/paper.bib
@@ -93,7 +93,7 @@
   note = {R package version 1.41.0,
     https://github.com/rstudio/reticulate},
   url = {https://rstudio.github.io/reticulate/},
-  doi = {https://doi.org/10.32614/CRAN.package.reticulate}
+  doi = {10.32614/CRAN.package.reticulate}
 }
 
 @software{louail_metabonaut_2025,
