### 2025-04-07 #8018 @jorainer
Potential reviewers could be:

- Florian Huber, author of the matchms Python library ( at [florian-huber](https://reviewers.joss.theoj.org/reviewers/1274)); also, *matchms* is also available in JOSS: https://joss.theoj.org/papers/10.21105/joss.02411
- Niek de Jonge, also developer in matchms (github handle: niekdejonge)
- Adriano Rutz ([Adriano Rutz](https://reviewers.joss.theoj.org/reviewers/3673))

or other reviewers from Mass Spectrometry field such as:

- [jspaezp](https://reviewers.joss.theoj.org/reviewers/1101)
- [yguitton](https://reviewers.joss.theoj.org/reviewers/1269)
- [biswapriyamisra](https://reviewers.joss.theoj.org/reviewers/89)
- [AnthonyOfSeattle](https://reviewers.joss.theoj.org/reviewers/1981)

Also, I fixed the missing DOI in the references in the latest.
