### 2025-04-14 #8018 @danielskatz
👋 @jspaezp & @yguitton & @biswapriyamisra & @AnthonyOfSeattle, would one or more of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
