### 2025-04-14 #8018 @danielskatz
> > 👋 ... would one or more of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> [@danielskatz](https://github.com/danielskatz) Would be happy to!

Thanks @Adafede - I will add you in our system, and will start the review once I get one or two more reviewers
