### 2025-04-14 #8018 @danielskatz
> I would be happy to help. I used to work with Wout Bittremieux several years ago, but if that is not a conflict I will join on.

@AnthonyOfSeattle - Can you look at our COI policy is https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy and see if you have a formal conflict? If so, if might also be possible to waive, if you can say a little more about how long ago this was.
