### 2025-04-14 #8018 @AnthonyOfSeattle
> > I would be happy to help. I used to work with Wout Bittremieux several years ago, but if that is not a conflict I will join on.
>
> [@AnthonyOfSeattle](https://github.com/AnthonyOfSeattle) - Can you look at our COI policy is https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy and see if you have a formal conflict? If so, if might also be possible to waive, if you can say a little more about how long ago this was.

I don't think I have a formal conflict according to that. I was affiliate of Wout's post-doc lab when we were both at the University of Washington 3-4 years ago. We had plenty of conversations and he helped me a great deal with a project for a graduate class. However, we do not have any publications together and I was not a major part of any of his projects.
