### 2025-04-15 #8018 @niekdejonge
@danielskatz I would be willing to, but I suspect there might be too much conflict of interest for the following reasons:

1. Published together with Helge Hecht in 2024 (https://jcheminf.biomedcentral.com/articles/10.1186/s13321-024-00878-1). We worked together on matchms and I still regularly work together with Helge on matchms and know him well.
2. Published together with Wout Bittemieux in 2024. (https://openreview.net/forum?id=AAo8zAShX3#discussion) This is a community paper and I did not work very closely with Wout on this project.
3. I am one of the main developers of matchms at the moment, which is wrapped for R in this work. This might be a benefit for the review, but also feels like a conflict of interest to me.

Let me know if these are not considered conflict of interests. In that case I am happy to do a review.
