### 2025-04-15 #8018 @danielskatz
@niekdejonge - Yes, the first is certainly a COI, the second might be, and I don't think the third actually is, but in any case, this is enough that we shouldn't have you review this. Thanks for responding and explaining the situation though.
