### 2025-04-17 #8018 @danielskatz
Let's go ahead and start.
