### 2025-04-18 cee5342
feat: backendInitialize to support data parameter
