### 2025-04-22 22d5f81
feat: add support for data parameter in backendInitialize()
