### 2025-04-22 c92a11e
ci: update to BioC 3.22
