### 2025-04-22 ebdad2d
ci: add codecov token
