### 2025-05-02 rformassspectrometry/SpectriPy#61 @Adafede
JOSS review
Dear @jorainer

I have been assigned to review your repository for https://github.com/openjournals/joss-reviews/issues/8070.

Please be ensured I highly appreciate a lot of points that are already available.
Per convenience, I will not list all the things that are already done well, but only the "issues" so this might give a way more negative impression than what my real thoughts are, my apologies for it already!

I will post the details in sub-issues and list them here gradually:
- [x] #62
- [x] #65
- [x] #66
- [x] #67

Extras I thought about but not necesarily to be included for the `JOSS review`:
- [x] #63
- [x] #64
- [ ] #68
