### 2025-05-02 rformassspectrometry/SpectriPy#62 @Adafede
[DOCUMENTATION] Title and what the package does
As per JOSS guidelines, a clear statement of need should be included, and this is the case.

While reading them at https://rformassspectrometry.github.io/SpectriPy/index.html, I got slightly confused as mainly the tile includes _less_ than what the package actually does.

See for example:
https://github.com/rformassspectrometry/SpectriPy/blob/ebdad2dddccebb2722aead655971f40b61dfc20e/DESCRIPTION#L2

I believe
https://github.com/rformassspectrometry/SpectriPy/blob/ebdad2dddccebb2722aead655971f40b61dfc20e/README.md?plain=1#L26-L27

Reflects what the package does much better. There are a few other places where the description are `matchms` focused and do not include `spectrum_utils`. I totally get this is probably because it evolved over time, making the different paragraphs match current version would probably show actual possibilities better.
