### 2025-05-02 rformassspectrometry/SpectriPy#65 @Adafede
[MANUSCRIPT] DESCRIPTION paragraph
The description paragraph in the article is very short and I believe including some parts (at least a short summary) of the external points directly in the paragraph would help.
Including the direct link https://rformassspectrometry.github.io/Metabonaut/articles/SpectriPy_tutorial_metabonaut.html would probably also ease the reader to land directly at the right place and not have to search.
