### 2025-05-02 rformassspectrometry/SpectriPy#66 @Adafede
[MANUSCRIPT] What about `pyteomics`/`pyOpenMS` functionalities?
The number of python packages included is already impressive, but what about `pyteomics`/`pyOpenMS`? It feels like they are major players in python MS pipelines and are mentioned in your summary but there are no functionalities to convert to these?

Maybe shortly mentioning if this is planned in the future or why not would help
