### 2025-05-02 rformassspectrometry/SpectriPy#67 @Adafede
[MANUSCRIPT] Add an actual example of what `SpectriPy` now allows
I find the plot just above https://rformassspectrometry.github.io/SpectriPy/articles/SpectriPy.html#summary **very** interesting and a nice illustration of which questions `SpectriPy` can help answering. Such comparisons are more complicated without it. I think including it in the manuscript would be a good idea. This would also help answering `how this software compares to other commonly-used packages`
