### 2025-05-02 #8070 @Adafede
Hi @jorainer, I believe I finished reviewing your submission, thank you very much for it, I appreciated it a lot!
You can find my comments above or in https://github.com/rformassspectrometry/SpectriPy/issues/61.
