### 2025-05-05 ac0d0b0
tests: add unit tests for zzz.R
