### 2025-05-05 #8070 @jorainer
@Adafede , thanks for the review! Quick feedback on the comment regarding the authors

> I think the list of authors is fine but could not find @ lgatto (minor contribution) in the authors. On the other side, I could not see any GitHub contribution from @ philouail directly on SpectriPy, the link to Metabonaut might not be straightforward to everyone.

@ philouail was working with me on the `MsBackendPy` class which involved a lot of offline discussions on the class design and concept. She then finally also reviewed the related PR ( https://github.com/rformassspectrometry/SpectriPy/pull/45 ). Regarding @ lgatto , the contribution to this particular repository were only minor (up to now), with commits to the README and DESCRIPTION files. If OK with you I would like to keep the list of authors as we have it now.
