### 2025-05-05 01b70b0
review: update title and improve description
- Address comments from @adafede from issue #62:
- Update the package title and description.
