### 2025-05-05 ef4d906
Update README.md
rm trailing slashes
