### 2025-05-05 358cd5c
improve description in devnotes.md
