### 2025-05-05 f6d9ce6
issue_66 edit in paper.md
review: [MANUSCRIPT] What about pyteomics/pyOpenMS functionalities? #66
- Address comment from @Adafede from issue #66:
- Explicit mention pyteomics/pyOpenMS in perspective section.
