### 2025-05-05 98869bb
issue_65 edit in paper.md
review: [MANUSCRIPT] DESCRIPTION paragraph #65
- Address comment from @Adafede from issue #65: 
- Explicit mention link to Metabonaut.
- Reiterate main uses of SpectriPy, instead of solely referring.
