### 2025-05-05 1cda1ce
issue_67 edit in paper.md
review: [MANUSCRIPT] Add an actual example of what SpectriPy now allows #67
- Address comment from @Adafede from issue #67:
- add example of use SpectriPy
- add comparison plot between 2 spectral similarity functions results
- mention advantage of SpectriPy to other software (ease of calculation and compare)
