### 2025-05-06 8a04b21
paper: small reformulation
