### 2025-05-06 rformassspectrometry/SpectriPy#62 @jorainer
Dear @Adafede we thank you for raising this issue. Indeed, we first implemented support for *matchms* and added *spectrum_utils* only later. We have now expanded the description and updated the title as suggested.

Please see our changes in PR #69 and let us know if this solves the issue - we would then merge the changes into the main branch.
