### 2025-05-06 ea43a8d
paper: add figure legend and move image
