### 2025-05-06 rformassspectrometry/SpectriPy#65 @jorainer
Thank you for this comment @Adafede ! We kept the different paragraphs and sections short because of the suggested word limit from JOSS. We have now expanded the description and added as suggested a direct link to the respective workflow in Metabonaut. You can find the changes in PR #70 - please let us know if it solves the issue and we will merge into the main branch.
