### 2025-05-06 rformassspectrometry/SpectriPy#66 @jorainer
Thank you for the suggestion @Adafede . We plan to support these libraries too in future and added this information to the perspective section. See PR #71 for the actual change in the wording. Please let us know if this addresses your comment and we will merge the changes into the main branch.
