### 2025-05-06 rformassspectrometry/SpectriPy#67 @jorainer
Thank you for the suggestion @Adafede ! We have now added an additional code chunk/example including the mentioned figure into the main text. The changes are in PR #72 - please let us know if this addresses your comment and we will merge the changes into the main branch.
