### 2025-05-06 #8070 @jorainer
Dear @Adafede , thank you so much for your thorough review - and for organizing the the comments/points as issues. We have now addressed all your issues and made a PR with suggested changes for each of them. We will then only merge these into the main branch once we have your OK. The mentioned PR and related issues are:

- [ ] https://github.com/rformassspectrometry/SpectriPy/pull/69 addressing https://github.com/rformassspectrometry/SpectriPy/issues/62

- [ ] https://github.com/rformassspectrometry/SpectriPy/pull/70 addressing https://github.com/rformassspectrometry/SpectriPy/issues/65

- [ ] https://github.com/rformassspectrometry/SpectriPy/pull/71 addressing https://github.com/rformassspectrometry/SpectriPy/issues/66

- [ ] https://github.com/rformassspectrometry/SpectriPy/pull/72 addressing https://github.com/rformassspectrometry/SpectriPy/issues/67
