### 2025-05-08 rformassspectrometry/SpectriPy#61 @Adafede
Closing it as all comments were addressed, `spec2vec` and `ms2deepscore` will remain as bonus 😉
