### 2025-05-12 rformassspectrometry/SpectriPy#73 @AnthonyOfSeattle
[DOCUMENTATION] Broken links on front page
Minor issue with links on the repo README and website frontpage. Because you have both of these, some links will 404 depending on whether you start from github.com or rformassspectrometry.github.io.

Some suspects:
https://github.com/rformassspectrometry/SpectriPy/blob/c823e95498936aeee8204463a92946f867a7c1c4/README.md?plain=1#L34

And
https://github.com/rformassspectrometry/SpectriPy/blob/c823e95498936aeee8204463a92946f867a7c1c4/README.md?plain=1#L74-L76
