### 2025-05-12 rformassspectrometry/SpectriPy#74 @AnthonyOfSeattle
[MANUSCRIPT] Clearer language around implementation
I had a hard time with my first couple reads of the following sentences:

https://github.com/rformassspectrometry/SpectriPy/blob/c823e95498936aeee8204463a92946f867a7c1c4/paper/paper.md?plain=1#L145-L150

In my opinion, this starts a bit too abstract. Instead, I would recommend starting by introducing the concrete example of how `SpectriPy` currently references python spectra before moving on. I think this will paint a clearer picture of what you mean by "backend class". Its likely that a quick rearrangement is all that you need. When you do move on to extensions, I think it would be interesting to see some brief specifics about the "different backend implementations" that you are hoping to implement. However, I think focusing on clarity is more important.
