### 2025-05-12 rformassspectrometry/SpectriPy#75 @AnthonyOfSeattle
[DOCUMENTATION] Detailed installation instructions for non-conda users
**NOT BLOCKING FOR PUBLICATION**

This request is very specific to my situation. The current miniconda installation instructions were very straight forward and I was able to create this Dockerfile to give me an environment for testing.

```
FROM rocker/r-ver:4.4

RUN R -e "install.packages('reticulate')"
RUN R -e "reticulate::install_miniconda()"

RUN R -e "install.packages('BiocManager')"
RUN R -e "BiocManager::install(version = '3.20')"

RUN R -e "install.packages('remotes')"
RUN R -e "BiocManager::install('RforMassSpectrometry/ProtGenerics', ask=F)"
RUN R -e "BiocManager::install('RforMassSpectrometry/Spectra', ask=F)"
RUN R -e "BiocManager::install('RforMassSpectrometry/SpectriPy', ask=F)"

RUN R -e "install.packages('testthat')"
```

However, I have a second weird situation. Due to commercial restrictions, I am unable to use miniconda at work. So I was looking for some detailed instructions on how to setup the package using `virtualenv` instead. I found the toggle between the two setups in the code chunk bellow,

https://github.com/rformassspectrometry/SpectriPy/blob/c823e95498936aeee8204463a92946f867a7c1c4/R/zzz.R#L68-L98

And I was able to find a reference to an the R variable that I needed to set here

https://github.com/rformassspectrometry/SpectriPy/blob/c823e95498936aeee8204463a92946f867a7c1c4/R/zzz.R#L55-L59

But I couldn't seem to get all the way through tests. Could these instructions be added to the Documentation?
