### 2025-05-13 549477a
docs: add link to Python configuration to the README (issue #75)
