### 2025-05-13 rformassspectrometry/SpectriPy#75 @jorainer
Thanks for pointing this out @AnthonyOfSeattle ! I've added a link to the respective documentation for the Python configuration to the README (see latest commit referenced above).

For the docker image we use for our [Metabonaut](https://github.com/RforMassSpectrometry/Metabonaut) resource we set

```
## Use SpectriPy with virtual env to avoid need to install miniconda
ENV SPECTRIPY_USE_CONDA="FALSE"
```

before installing *SpectriPy*. See also the respective lines in the Docker file: https://github.com/rformassspectrometry/Metabonaut/blob/devel/Dockerfile#L18-L19

Let me know if your problem still persists.
