### 2025-05-13 cbf2710
Reformulate and remove backend description
