### 2025-05-13 a18c141
docs: fix links in the README (issue #73)
