### 2025-05-13 rformassspectrometry/SpectriPy#73 @jorainer
Thanks @AnthonyOfSeattle ! I fixed now the links. They point now to the correct documents - and should work both on github and the rendered package page.
