### 2025-05-13 #8070 @danielskatz
@AnthonyOfSeattle - I see that there's one unchecked item in [your checklist](https://github.com/openjournals/joss-reviews/issues/8070#issuecomment-2827772953). Do you think there's something that @jorainer should be doing to address this?
