### 2025-05-13 #8070 @AnthonyOfSeattle
Good morning all,

Thanks for your patience as I balancing some other work with this review.

The authors have developed a very nice linking package between Mass Spectrometry software development in R and Python. I agree with the authors' Statement of Need that this has the potential to reduce the need for reinventing the wheel in both languages and would allow developers to leverage the strengths of both. In its current state, the package addresses a clear use case of translating spectrum objects between languages and allowing R developers to take advantage of `matchms`. Looking through the paper and issues, it looks like there are plans to build out the package further and make it even more functional. The documentation around individual methods is thorough and I had no problem with setup for the typical setup scenario.

I have only a couple minor tickets, one of which was addressed and one which I would like addressed.

- [x] https://github.com/rformassspectrometry/SpectriPy/issues/73

- [ ] https://github.com/rformassspectrometry/SpectriPy/issues/74

The authors were also very responsive in helping me with a specific setup case that I did not consider blocking for publication. I will need time to try their suggestions but I am thankful for their feedback.

- https://github.com/rformassspectrometry/SpectriPy/issues/75
