### 2025-05-14 5bd257d
reformulate example description
