+++ paper/paper.md
@@ -144,7 +144,8 @@ conversion, and any required renaming and reformatting of spectra metadata, such
 as MS level, retention times, or any other arbitrary metadata available in the
 MS data object. An example combined R-Python data analysis workflow, which can
 be realized e.g. using the Quarto system is provided in the following code
-snippets. In the Python code block below, MS data are imported and processed.
+snippets. In this particular example we start the analysis in Python, loading
+and processing the MS data with functions from the *matchms* library.
 
 ```python
 #' Python session:
@@ -182,7 +183,6 @@ peaksData(sps[1])
 The use of the `MsBackendPy` enables thus seamless and, compared to the
 alternative `pyspec_to_rspec()`, more memory-efficient integration of Python MS
 data objects into R for powerful cross-language analysis workflows.
-
 
 ## Integrated functionality from the *matchms* Python library
 
