### 2025-05-14 rformassspectrometry/SpectriPy#73 @jorainer
Perfect! Thanks for the quick feedback! I leave it to you to close the issue.
