### 2025-05-14 rformassspectrometry/SpectriPy#74 @jorainer
Thanks for the suggestions @AnthonyOfSeattle ! Indeed, the information on the concept of the backends was a bit too abstract - we reformulated it now, also removing any information that is not important or relevant for *SpectriPy* (whether the backend in memory-saving etc). You can see the new text in the PR #76 . Please have a look and let us know if you are OK with it (in which case we will merge the PR into the main branch) or if we need to still adapt.
