### 2025-05-15 rformassspectrometry/SpectriPy#74 @AnthonyOfSeattle
I like this a lot better. Feel free to merge and I will close.
