### 2025-05-16 rformassspectrometry/SpectriPy#74 @jorainer
I merged it now to the main branch. Thanks for the feedback @AnthonyOfSeattle !
