### 2025-05-16 #8070 @danielskatz
@jorainer - my suggested edits are in https://github.com/rformassspectrometry/SpectriPy/pull/77 - please merge this, or let me know what you disagree with.

After that, please:
