### 2025-05-19 5e6ae83
Create CITATION.cff
