### 2025-05-19 #8070 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
